day 11- renamed methods

+++ src/Day11/Part1and2/Coordinates.java
@@ -1,6 +1,6 @@
 package Day11.Part1and2;
 
-public record Coordinates(long row, long column) {
+public record Coordinates(int row, int column) {
